fixed chucker inspection

+++ app/src/main/java/com/example/weatherforecast/data/models/network/Wind.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 
 data class Wind(
     @SerializedName("deg")
-    val deg: Int,
+    val deg: Double,
     @SerializedName("speed")
-    val speed: Int
+    val speed: Double
 )
